fix styles
