+++ src/components/ColorBlock/ColorBlock.scss
@@ -1,0 +1,28 @@
+@import "../../Globals.scss";
+
+.color-block {
+  display: grid;
+  cursor: pointer;
+  border: none;
+  background-color: #fff;
+  width: 100%;
+  font-family: Verdana;
+  font-weight: 400;
+  font-size: 1.1rem;
+  color: filter;
+  .edit-prompt {
+    grid-column: 1;
+    width: 80px;
+    opacity: 0;
+    &:hover {
+        opacity: 1;
+    }
+
+  }
+  .copy-prompt {
+      padding-left: 26px;
+    text-align: left;
+    grid-column: 2;
+    width: 220px;
+  }
+}
